RegularlySampledFormat

The RegularlySampledFormat package provides an interface for working with RSF files.

The RSF file format is used in the Madagascar open-source package for geophysical data analysis.

RegularlySampledFormat.add!

RegularlySampledFormat.enter!

RegularlySampledFormat.getbool

RegularlySampledFormat.getbools

RegularlySampledFormat.getfloat

RegularlySampledFormat.getfloats

RegularlySampledFormat.getint

RegularlySampledFormat.getints

RegularlySampledFormat.getpar

RegularlySampledFormat.getpars

RegularlySampledFormat.getstring

RegularlySampledFormat.input!

RegularlySampledFormat.output

RegularlySampledFormat.string!

RegularlySampledFormat.add! — Method

add!(t, keyval)

Enter a key-value pair into the symbolic table `t` from a string of the form "key=val".

RegularlySampledFormat.enter! — Method

enter!(t, key, val)

Enter a key-value pair into the symbolic table `t`.

RegularlySampledFormat.getbool — Method

getbool(t, key)

Extract a boolean parameter from the symbolic table `t` for a given `key`.

RegularlySampledFormat.getbools — Method

getbools(t, key, n)

Extract an array of boolean parameters from the symbolic table `t` for a given `key`. 
The array will have length `n`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned.

RegularlySampledFormat.getfloat — Method

getfloat(t, key)

Extract a float parameter from the symbolic table `t` for a given `key`.

RegularlySampledFormat.getfloats — Method

getfloats(t, key, n)

Extract an array of float parameters from the symbolic table `t` for a given `key`. 
The array will have length `n`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned.

RegularlySampledFormat.getint — Method

getint(t, key)

Extract an integer parameter from the symbolic table `t` for a given `key`.

RegularlySampledFormat.getints — Method

getints(t, key, n)

Extract an array of integer parameters from the symbolic table `t` for a given `key`. 
The array will have length `n`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned.

RegularlySampledFormat.getpar — Method

add!(t, keyval)

Enter a key-value pair into the symbolic table `t` from a string of the form "key=val".

RegularlySampledFormat.getpars — Method

getpars(t, key, n, T)

Extract an array of parameters from the symbolic table `t` for a given `key`. 
The array will have length `n` and will be of type `T`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned.

RegularlySampledFormat.getstring — Method

getstring(t, key)

Extract a string parameter from the symbolic table `t` for a given `key`.

RegularlySampledFormat.input! — Function

input!(t, filep)

Input key-value pairs into the symbolic table `t` from a file stream `filep`.

RegularlySampledFormat.output — Method

output(t, filep)

Output the contents of the symbolic table `t` to a file stream `filep`.

RegularlySampledFormat.string! — Method

string!(t, string)

Enter key-value pairs into the symbolic table `t` from a string of the form "key1=val1 key2=val2 ...".