RegularlySampledFormat

The RegularlySampledFormat package provides an interface for working with RSF files.

The RSF file format is used in the Madagascar open-source package for geophysical data analysis.

RegularlySampledFormat.Input

RegularlySampledFormat.Output

RegularlySampledFormat._RSF

RegularlySampledFormat.Datapath

RegularlySampledFormat.Temp

RegularlySampledFormat.add!

RegularlySampledFormat.bytes

RegularlySampledFormat.enter!

RegularlySampledFormat.fflush!

RegularlySampledFormat.fileflush!

RegularlySampledFormat.floatread!

RegularlySampledFormat.floatwrite

RegularlySampledFormat.getbool

RegularlySampledFormat.getbool

RegularlySampledFormat.getbools

RegularlySampledFormat.getbools

RegularlySampledFormat.getfilename

RegularlySampledFormat.getfloat

RegularlySampledFormat.getfloat

RegularlySampledFormat.getfloats

RegularlySampledFormat.getfloats

RegularlySampledFormat.getform

RegularlySampledFormat.getint

RegularlySampledFormat.getint

RegularlySampledFormat.getints

RegularlySampledFormat.getints

RegularlySampledFormat.getpar

RegularlySampledFormat.getpar

RegularlySampledFormat.getpars

RegularlySampledFormat.getpars

RegularlySampledFormat.getprog

RegularlySampledFormat.getstring

RegularlySampledFormat.getstring

RegularlySampledFormat.getstring

RegularlySampledFormat.gettype

RegularlySampledFormat.input!

RegularlySampledFormat.intread!

RegularlySampledFormat.intwrite

RegularlySampledFormat.output

RegularlySampledFormat.putfloat!

RegularlySampledFormat.putfloats!

RegularlySampledFormat.putint!

RegularlySampledFormat.putints!

RegularlySampledFormat.putstring!

RegularlySampledFormat.read!

RegularlySampledFormat.setform!

RegularlySampledFormat.setformat!

RegularlySampledFormat.settype!

RegularlySampledFormat.string!

RegularlySampledFormat.tell

RegularlySampledFormat.ucharwrite

RegularlySampledFormat.write

RegularlySampledFormat.Input — Method

Input(tag::String) -> Input

Create an instance of the Input struct for reading from an RSF file specified by tag.

RegularlySampledFormat.Output — Method

Output(tag::String) -> Output

Create an instance of the Output struct for writing to an RSF file specified by tag.

RegularlySampledFormat._RSF — Type

_RSF(inp::Bool, tag=nothing) -> _RSF

Create an instance of the `_RSF` struct for reading or writing RSF files. 
If `inp` is true, the function reads from an input RSF file specified by `tag`. 
If `inp` is false, it prepares for writing to an output RSF file specified by `tag`.

RegularlySampledFormat.Datapath — Method

Datapath() -> String

Return the path to the data directory. The path is determined by checking the following sources in order of priority:
1. The environment variable `DATAPATH`.
2. A file named `.datapath` in the current working directory.
3. A file named `.datapath` in the user's home directory.
4. If none of the above sources provide a path, the function defaults to returning
the current working directory (`"./"`).

RegularlySampledFormat.Temp — Method

Temp() -> (String, IO)

Create a temporary data path. The temporary data path is determined by checking the following sources in order of priority:
1. The environment variable `TMPDATAPATH`.
2. The result of `Datapath()` function.

Returns a tuple containing the path to the temporary file and an IO object for writing to the file.

RegularlySampledFormat.add! — Method

add!(t, keyval)

Enter a key-value pair into the symbolic table `t` from a string of the form "key=val".

RegularlySampledFormat.bytes — Method

bytes(rsf::_RSF) -> Int

Return the size in bytes of the RSF file represented by the `_RSF` struct. 
If the data name is "stdin", it returns -1.

RegularlySampledFormat.enter! — Method

enter!(t, key, val)

Enter a key-value pair into the symbolic table `t`.

RegularlySampledFormat.fflush! — Method

fflush!(rsf::_RSF)

Flush the contents of the RSF file represented by the `_RSF` struct to disk.

RegularlySampledFormat.fileflush! — Method

fileflush!(rsf::_RSF, src::_RSF)

Flush the contents of the RSF file represented by the `_RSF` struct to disk. 
If `src` is provided, it also flushes the contents of the source RSF file to the destination.

RegularlySampledFormat.floatread! — Method

floatwrite(rsf::_RSF, arr)

Write an array of floats (Float32) to the RSF file represented by the `_RSF` struct.

RegularlySampledFormat.floatwrite — Method

floatwrite(rsf::_RSF, arr)

Write an array of floats (Float32) to the RSF file represented by the `_RSF` struct.

RegularlySampledFormat.getbool — Function

getbool(key, default)

Extract a boolean parameter from the parameter table `par` for a given `key`. 
If the key is not found, return the `default` value.

RegularlySampledFormat.getbool — Method

getbool(t, key)

Extract a boolean parameter from the symbolic table `t` for a given `key`.

RegularlySampledFormat.getbools — Function

getbools(key, n, default)

Extract an array of boolean parameters from the parameter table `par` for a given `key`. 
The array will have length `n`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned. If the key is not found, return the `default` value.

RegularlySampledFormat.getbools — Method

getbools(t, key, n)

Extract an array of boolean parameters from the symbolic table `t` for a given `key`. 
The array will have length `n`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned.

RegularlySampledFormat.getfilename — Method

getfilename(stream::IOStream) -> Union{String, Symbol}

Find the name of the file associated with the given IO stream. 
If the stream is not associated with a file, returns `:none`.

RegularlySampledFormat.getfloat — Function

getfloat(key, default)

Extract a float parameter from the parameter table `par` for a given `key`. 
If the key is not found, return the `default` value.

RegularlySampledFormat.getfloat — Method

getfloat(t, key)

Extract a float parameter from the symbolic table `t` for a given `key`.

RegularlySampledFormat.getfloats — Function

getfloats(key, n, default)

Extract an array of float parameters from the parameter table `par` for a given `key`. 
The array will have length `n`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned. If the key is not found, return the `default` value.

RegularlySampledFormat.getfloats — Method

getfloats(t, key, n)

Extract an array of float parameters from the symbolic table `t` for a given `key`. 
The array will have length `n`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned.

RegularlySampledFormat.getform — Method

getform(rsf::_RSF) -> String

Return the data format of the RSF file represented by the `_RSF` struct.

RegularlySampledFormat.getint — Function

getint(key, default)

Extract an integer parameter from the parameter table `par` for a given `key`. 
If the key is not found, return the `default` value.

RegularlySampledFormat.getint — Method

getint(t, key)

Extract an integer parameter from the symbolic table `t` for a given `key`.

RegularlySampledFormat.getints — Function

getints(key, n, default)

Extract an array of integer parameters from the parameter table `par` for a given `key`. 
The array will have length `n`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned. If the key is not found, return the `default` value.

RegularlySampledFormat.getints — Method

getints(t, key, n)

Extract an array of integer parameters from the symbolic table `t` for a given `key`. 
The array will have length `n`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned.

RegularlySampledFormat.getpar — Function

getpar(p, key, T, default)

Extract a parameter of type `T` from the parameter table `p` for a given `key`. 
If the key is not found, return the `default` value.

RegularlySampledFormat.getpar — Method

getpar(t, key, T)

Extract a parameter of type `T` from the symbolic table `t` for a given `key`.

RegularlySampledFormat.getpars — Function

getpars(p, key, n, T, default)

Extract an array of parameters of type `T` from the parameter table `p` for a given `key`. 
The array will have length `n`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned. If the key is not found, return the `default` value.

RegularlySampledFormat.getpars — Method

getpars(t, key, n, T)

Extract an array of parameters from the symbolic table `t` for a given `key`. 
The array will have length `n` and will be of type `T`. If the number of values in the table is less than `n`, the last value will be repeated to fill the array. If the number of values is greater than `n`, only the first `n` values will be returned.

RegularlySampledFormat.getprog — Method

getprog(p)

Extract the program name from the parameter table `p`.

RegularlySampledFormat.getstring — Function

getstring(key, default)

Extract a string parameter from the parameter table `par` for a given `key`. 
If the key is not found, return the `default` value.

RegularlySampledFormat.getstring — Function

getstring(rsf::_RSF, key::String, default=:none) -> String

Retrieve the string value associated with the specified `key` in the RSF file represented by the `_RSF` struct.
If the key is not found, return the `default` value.

RegularlySampledFormat.getstring — Method

getstring(t, key)

Extract a string parameter from the symbolic table `t` for a given `key`.

RegularlySampledFormat.gettype — Method

gettype(rsf::_RSF) -> DataType

Return the data type of the RSF file represented by the `_RSF` struct.

RegularlySampledFormat.input! — Function

input!(t, filep)

Input key-value pairs into the symbolic table `t` from a file stream `filep`.

RegularlySampledFormat.intread! — Method

intread!(rsf::_RSF, arr)  

Read an array of integers (Int32) from the RSF file represented by the `_RSF` struct into the provided `arr`.

RegularlySampledFormat.intwrite — Method

intwrite(rsf::_RSF, arr)

Write an array of integers (Int32) to the RSF file represented by the `_RSF` struct.

RegularlySampledFormat.output — Method

output(t, filep)

Output the contents of the symbolic table `t` to a file stream `filep`.

RegularlySampledFormat.putfloat! — Method

putfloats!(rsf::_RSF,key::String,par::Array{Float32},n::Int)

Write an array of float parameters to the RSF file represented by the `_RSF` struct. 
The parameters are associated with the specified `key`.

RegularlySampledFormat.putfloats! — Method

putfloats!(rsf::_RSF,key::String,par::Array{Float32},n::Int)

Write an array of float parameters to the RSF file represented by the `_RSF` struct. 
The parameters are associated with the specified `key`.

RegularlySampledFormat.putint! — Method

putint!(rsf::_RSF, key::String, par::Int)

Write an integer parameter to the RSF file represented by the `_RSF` struct. 
The parameter is associated with the specified `key`.

RegularlySampledFormat.putints! — Method

putfloat!(rsf::_RSF, key::String, par::Float)

Write a float parameter to the RSF file represented by the `_RSF` struct. 
The parameter is associated with the specified `key`.

RegularlySampledFormat.putstring! — Method

putstring!(rsf::_RSF, key::String, par::String)

Write a string parameter to the RSF file represented by the `_RSF` struct. 
The parameter is associated with the specified `key`.

RegularlySampledFormat.read! — Method

read!(inp::Input, data::Array)

Read data from the RSF file represented by the `Input` struct into the provided `data` array.

RegularlySampledFormat.setform! — Method

setform!(rsf::_RSF, form::String)

Set the data format of the RSF file represented by the `_RSF` struct to `form`. 
If `form` is "ascii", it also sets the appropriate formats for ASCII output.

RegularlySampledFormat.setformat! — Method

setformat!(rsf::_RSF, dataformat::String)

Set the data type and format of the RSF file represented by the `_RSF` struct based on the provided `dataformat` string. 
The function determines the appropriate data type and format (ASCII, XDR, or native) based on the contents of `dataformat`.

RegularlySampledFormat.settype! — Method

settype!(rsf::_RSF, T::DataType)

Set the data type of the RSF file represented by the `_RSF` struct to `T`.

RegularlySampledFormat.string! — Method

string!(t, string)

Enter key-value pairs into the symbolic table `t` from a string of the form "key1=val1 key2=val2 ...".

RegularlySampledFormat.tell — Method

tell(rsf::_RSF) -> Int

Return the current position in the RSF file represented by the `_RSF` struct.

RegularlySampledFormat.ucharwrite — Method

ucharwrite(rsf::_RSF, arr)

Write an array of unsigned characters (UInt8) to the RSF file represented by the `_RSF` struct.

RegularlySampledFormat.write — Method

write(out::Output, data::Array)

Write data from the provided `data` array to the RSF file represented by the `Output` struct.